+++ changelog.md
@@ -1,3 +1,4 @@
 ## 0.1.0
 
 - Initial public repository + NVDA add-on build scaffolding.
+- Expanded command support (ROM, punctuation/spell/hyper/phonetic, mark-space ratio) and NVDA speech commands (break/prosody).